--- domain/src/main/java/woowacourse/paint/repository/PaintColorRepository.kt
@@ -1,10 +1,0 @@
-package woowacourse.paint.repository
-
-import woowacourse.paint.model.PaintColor
-
-interface PaintColorRepository {
-
-    fun getPaintColor(): PaintColor
-    fun setPaintColor(paintColor: PaintColor)
-    fun getAllPaintColors(): List<PaintColor>
-}
